+++ src/main/resources/release_notes.md
@@ -45,6 +45,7 @@
 ### Youth
 * Increase effect of youth friendly match training (#1950)
 * Modify youth module layout (#1449)
+* Set initial youth module view layout (#1558)
 
 ### Misc
 * Add a Linux-friendly default theme, called “Gnome.”
@@ -56,9 +57,10 @@
 Reports by Contributors - September 24, 2023 - December 07, 2023
 
 * Sebastien 157
+* \_KOHb\_ 22
 * Georgi 10
 * Lidegand 3
 * sich 3
 * wsbrenk 3
 
-Total 176
+Total 198